feat: add find product by id
